tests/wpt/tests/svg/animations/repeat-event.svg

<svg xmlns="http://www.w3.org/2000/svg">
  <title>Animation triggers on 'repeatEvent'</title>
  <script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharness.js"/>
  <script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharnessreport.js"/>
  <rect width="0" height="100" fill="green">
    <animate id="anim" attributeName="visibility" to="visible" begin="0s" dur="1s" repeatCount="2" fill="freeze"/>
    <set attributeName="width" to="100" begin="anim.repeatEvent"/>
  </rect>
  <script>
    async_test(t => {
        document.documentElement.setCurrentTime(0.99);

        let set = document.querySelector('set');
        set.addEventListener('beginEvent', t.step_func_done(function() {
            assert_equals(set.parentNode.getBBox().width, 100);
        }));
    });
  </script>
</svg>
